+++ R/Appender.R
@@ -139,6 +139,10 @@ Appender <- R6::R6Class(
 #' **crayon** installed log levels will be coloured by default
 #' (but you can modify this behaviour by passing a custom [Layout]).
 #'
+#' @param connection A connection or a `character` scalar. See the `file`
+#'   argument of [cat()] for more details. Defaults to [stdout()], except
+#'   inside \pkg{knitr} rendering processes where it defaults to [stderr()].
+#'
 #' @family Appenders
 #' @seealso [LayoutFormat]
 #' @export
@@ -168,24 +172,50 @@ AppenderConsole <- R6::R6Class(
         timestamp_fmt = "%H:%M:%OS3",
         colors = getOption("lgr.colors", list())
       ),
-      filters = NULL
+      filters = NULL,
+      connection = NULL
     ){
       self$set_threshold(threshold)
       self$set_layout(layout)
       self$set_filters(filters)
+      if (is.null(connection)) {
+        connection <- default_console_connection()
+      }
+      self$set_connection(connection)
     },
 
     append = function(event){
-      cat(private$.layout$format_event(event), sep = "\n")
+      cat(private$.layout$format_event(event), sep = "\n", file = get(".connection", private))
+    },
+
+    set_connection = function(connection){
+      assert(
+        inherits(connection, "connection") || is_scalar_character(connection)
+      )
+
+      private[[".connection"]] <- connection
     }
   ),
 
   active = list(
-    destination = function() "console"
+    destination = function() "console",
+    connection = function() private[[".connection"]]
+  ),
+
+  private = list(
+    .connection = NULL
   )
 )
 
+default_console_connection <- function() {
+  # In knitr, default to using stderr where log messages can be better handled
+  if (isTRUE(getOption("knitr.in.progress", FALSE))) {
+    return(stderr())
+  }
 
+  # has to be this, not a call to `stdout()`
+  ""
+}
 
 
 # AppenderFile ------------------------------------------------------------